refactor: split subagent.ts into focused modules
Split the 1,071-line monolith into a subagent/ directory:

- subagent/types.ts    — shared interfaces, constants, pure type utilities
- subagent/format.ts   — formatTokens, formatUsageStats, formatToolCall, aggregateUsage
- subagent/engine.ts   — BUILTIN_TOOLS, resolveTools, mapWithConcurrencyLimit, runSingleAgent
- subagent/render.ts   — renderSubagentCall, renderSubagentResult (extracted from registerTool)
- subagent/index.ts    — extension factory + JSON schema + re-exports for compat

subagent.ts becomes a one-liner re-export barrel so all existing callers
(factories.ts, subagent.test.ts) continue to work without changes.

No logic changes — pure reorganization. All 290 tests pass.
